Conversions refactor (#102)
Conversion code is now stored inside forcing module. Conversion is considered a sub-type of a forcing so that output filename is `<forcing_name>.<conversion_target_name>.nc`. Also introduces the concept of "conversion defaults" which
are input definition files for conversions which contain "sensible defaults" that can be used as starting point when setting up forcing conversion.

Also includes fixes for list serializion in output attrs and other minor tweaks (
nicer path for existing input defn, allow trajectory creation while domain downloads pending)
