+++ lagtraj/forcings/__init__.py
@@ -1,5 +1,7 @@
 from collections import namedtuple
-from ..utils.interpolation.levels import ForcingLevelsDefinition  # noqa
+from ..utils.interpolation.levels import (  # noqa
+    LevelsDefinition as ForcingLevelsDefinition,
+)
 from .profile_calculation import ForcingSamplingDefinition  # noqa
 from .. import build_data_path
 from ..input_definitions.examples import LAGTRAJ_EXAMPLES_PATH_PREFIX
@@ -26,13 +28,26 @@ ForcingDefinition = namedtuple(
 )
 
 
-def build_forcing_data_path(root_data_path, forcing_name):
+def build_forcing_data_path(root_data_path, forcing_name, conversion_name=None):
+    """
+    Create the output path for a forcing with name `forcing_name` in the path
+    `root_data_path`, optionally targetting a model by name `conversion_name`
+    """
     # we need to strip the `lagtraj://` prefix before we construct the path
     # since the data is stored locally
     if forcing_name.startswith(LAGTRAJ_EXAMPLES_PATH_PREFIX):
         forcing_name = forcing_name[len(LAGTRAJ_EXAMPLES_PATH_PREFIX) :]
 
-    filename = f"{forcing_name}.nc"
+    if conversion_name is not None and conversion_name.startswith(
+        LAGTRAJ_EXAMPLES_PATH_PREFIX
+    ):
+        conversion_name = conversion_name[len(LAGTRAJ_EXAMPLES_PATH_PREFIX) :]
+
+    if conversion_name is None:
+        filename = f"{forcing_name}.nc"
+    else:
+        filename = f"{forcing_name}.{conversion_name}.nc"
+
     return build_data_path(root_data_path=root_data_path, data_type="forcing") / (
-        forcing_name + ".nc"
+        filename
     )
