+++ lagtraj/utils/validation.py
@@ -21,7 +21,9 @@ def validate_trajectory(ds_traj):
             " fields: {}".format(", ".join(missing_fields))
         )
 
-    required_attrs = ["name", "domain_name", "trajectory_type"]
+    required_attrs = ["name", "trajectory_type"]
+    if ds_traj.attrs.get("trajectory_type") != "linear":
+        required_attrs += ["domain_name"]
     missing_attrs = list(filter(lambda f: f not in ds_traj.attrs, required_attrs))
 
     if len(missing_attrs) > 0:
