+++ tests/test_input_definition_examples.py
@@ -7,13 +7,15 @@ from lagtraj import DATA_TYPE_PLURAL, DEFAULT_ROOT_DATA_PATH
 from lagtraj.domain import INPUT_REQUIRED_FIELDS as DOMAIN_REQUIRED_FIELDS
 from lagtraj.forcings import INPUT_REQUIRED_FIELDS as FORCING_REQUIRED_FIELDS
 from lagtraj.trajectory import INPUT_REQUIRED_FIELDS as TRAJECTORY_REQUIRED_FIELDS
-from lagtraj.conversion import INPUT_REQUIRED_FIELDS as CONVERSION_REQUIRED_FIELDS
+from lagtraj.forcings.conversion import (
+    INPUT_REQUIRED_FIELDS as CONVERSION_REQUIRED_FIELDS,
+)
 
 INPUT_TYPES = {
     "forcing": FORCING_REQUIRED_FIELDS,
     "domain": DOMAIN_REQUIRED_FIELDS,
     "trajectory": TRAJECTORY_REQUIRED_FIELDS,
-    "conversion": CONVERSION_REQUIRED_FIELDS,
+    "forcing_conversion": CONVERSION_REQUIRED_FIELDS,
 }
 
 
@@ -50,10 +52,6 @@ def test_load_example(input_example):
     i = input_example.index("/")
     input_type_plural = input_example[:i]
     input_name = f"lagtraj://{input_example[i + 1 :]}"
-
-    if input_type_plural == "forcings_conversion":
-        # TODO: implement tests for the forcing conversions bundled with lagtraj
-        return
 
     input_type = None
     for k, v in DATA_TYPE_PLURAL.items():
